Do not change the CouchDB views (#3844)
In commit ce5f1b0219e189203954dddaef805f008bd4c217, we tried to optimize
the counting of files by using a CouchDB view. We changed the reduce
property, and I thought CouchDB won't recalculate the whole view. But it
does, and it is too heavy for our servers. So, let's use the view as it
is today, even if it needs more code and is a bit less optimized.

+++ model/sharing/sharing.go
@@ -59,7 +59,7 @@ type Sharing struct {
 	PreviewPath string    `json:"preview_path,omitempty"`
 	CreatedAt   time.Time `json:"created_at"`
 	UpdatedAt   time.Time `json:"updated_at"`
-	NbFiles     int64     `json:"initial_number_of_files_to_sync,omitempty"`
+	NbFiles     int       `json:"initial_number_of_files_to_sync,omitempty"`
 	Initial     bool      `json:"initial_sync,omitempty"`
 	ShortcutID  string    `json:"shortcut_id,omitempty"`
 	MovedFrom   string    `json:"moved_from,omitempty"`
